refactor(fs): extract realpath + resolveSymlinks into symlink-resolver.ts
